server/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyodbc==5.3.0
python-dotenv==1.2.1
Werkzeug==3.1.5
SQLAlchemy==2.0.46
marshmallow==3.26.2
flask-smorest==0.46.2
google-cloud-optimization==1.13.0
requests==2.31.0
googlemaps==4.6.0
